pyproject.toml
[tool.poetry]
name = "tibia-ipchanger"
version = "0.1.0"
description = "Command-line IP changer for Tibia clients (Linux only)"
authors = ["Ranieri Althoff <ranisalt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pyelftools = "^0.28"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
ipchanger = "ipchanger"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
